requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Caching
Werkzeug
requests
scikit-learn
pandas
joblib
xgboost